refactor: 에러 핸들링 유틸 파일 분리
