+++ AND-SOPT-iOS/AND-SOPT-iOS/Network/API/Users/LoginService.swift
@@ -42,39 +42,9 @@ class LoginService {
                     completion(.failure(.unknownError))
                 }
             case .failure:
-                let error = self.handleStatusCode(statusCode, data: data)
+                let error = NetworkUtils.handleStatusCode(statusCode, data: data)
                 completion(.failure(error))
             }
         }
-    }
-    func handleStatusCode(
-        _ statusCode: Int,
-        data: Data
-    ) -> NetworkError {
-        let errorCode = decodeError(data: data)
-        print("Status Code: \(statusCode), Error Code: \(errorCode)")
-        
-        switch (statusCode, errorCode) {
-        case (400, "00"):
-            return .invalidRequest
-        case (400, "01"):
-            return .expressionError
-        case (404, ""):
-            return .invalidURL
-        case (409, "00"):
-            return .duplicateError
-        case (500, ""):
-            return .serverError
-        default:
-            return .unknownError
-        }
-    }
-    
-    func decodeError(data: Data) -> String {
-        guard let errorResponse = try? JSONDecoder().decode(
-            ErrorResponse.self,
-            from: data
-        ) else { return "" }
-        return errorResponse.code
     }
 }
