+++ AND-SOPT-iOS/AND-SOPT-iOS/Network/NetworkUtils.swift
@@ -1,0 +1,41 @@
+//
+//  NetworkUtils.swift
+//  AND-SOPT-iOS
+//
+//  Created by 김예지 on 11/8/24.
+//
+
+import Foundation
+
+struct NetworkUtils {
+    static func handleStatusCode(
+        _ statusCode: Int,
+        data: Data
+    ) -> NetworkError {
+        let errorCode = decodeError(data: data)
+        print("Status Code: \(statusCode), Error Code: \(errorCode)")
+        
+        switch (statusCode, errorCode) {
+        case (400, "00"):
+            return .invalidRequest
+        case (400, "01"):
+            return .expressionError
+        case (404, ""):
+            return .invalidURL
+        case (409, "00"):
+            return .duplicateError
+        case (500, ""):
+            return .serverError
+        default:
+            return .unknownError
+        }
+    }
+    
+    static func decodeError(data: Data) -> String {
+        guard let errorResponse = try? JSONDecoder().decode(
+            ErrorResponse.self,
+            from: data
+        ) else { return "" }
+        return errorResponse.code
+    }
+}
